% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructGroupList.R
\name{constructGroupList}
\alias{constructGroupList}
\title{Internal function to construct a list of plotting parameters.}
\usage{
constructGroupList(items, population, include_nonsampled = TRUE,
  var_name = NULL, col = NULL, selected_point_cex = 1/2,
  nonselected_col = "gray", nonselected_point_cex = 1/4,
  nonselected_alpha = 1/3)
}
\arguments{
\item{items}{a character vector or list of character vectors representing 
groups of items used to index population}

\item{population}{a matrix or dataframe containing describing each item in
the population on the dimensions to be plotted}

\item{include_nonsampled}{logical deonting whether to construct a group for
all items that haven't been selected}

\item{var_name}{string or vector of length 2 (c(y,x)) with names of variables
of interest}

\item{col}{string or vector of colours for selected group(s) of items}

\item{selected_point_cex}{cex for scatterplot points for selected items}

\item{nonselected_col}{colour for scatterplot points for nonselected items}

\item{nonselected_point_cex}{cex for scatterplot points for nonselected items}

\item{nonselected_alpha}{transparency for scatterplot points for nonselected
items}
}
\description{
Internal function to construct a list of plotting parameters.
}
