% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDensityCurve.R
\name{drawDensityCurve}
\alias{drawDensityCurve}
\title{Internal function to draw density curve.}
\usage{
drawDensityCurve(d, col)
}
\arguments{
\item{d}{a density object}

\item{col}{line colour}
}
\description{
Internal function to draw density curve.
}
