% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateSubset.R
\name{evaluateSubset}
\alias{evaluateSubset}
\title{Computes distance-based descriptive statistics for a sample of items.}
\usage{
evaluateSubset(indices, distance_mat)
}
\arguments{
\item{indices}{a numeric or logical vector of the sample of items}

\item{distance_matrix}{a matrix containing inter-item distances for the
entire population}
}
\value{
a vector of descriptive statistics for the sample
}
\description{
Computes distance-based descriptive statistics for a sample of items.
}
