% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnessFunction.R
\name{fitnessFunction}
\alias{fitnessFunction}
\title{Function to sample items with a genetic algorithm maximizing some measure of 
fitness (e.g., everage inter-item dissimilarity) as measured by an inter-item
distance matrix.}
\usage{
fitnessFunction(x, sample_size, distance_mat,
  lower_tri_function = sumHeight, required_items = NULL)
}
\arguments{
\item{x}{a binary vector of sampled items}

\item{sample_size}{the number of items to sample}

\item{distance_mat}{a population inter-item distance matrix}

\item{required_items}{character vector of any items required to be included
in the solution}

\item{lower_tri_funciton}{the specific function to assess fitness of the 
lower triangle of the sample distance matrix}
}
\value{
a value representing the fitness of the current sample
}
\description{
Function to sample items with a genetic algorithm maximizing some measure of 
fitness (e.g., everage inter-item dissimilarity) as measured by an inter-item
distance matrix.
}
