% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt}
\alias{fmt}
\title{Internal function to format numbers for printing and plotting.}
\usage{
fmt(x, digits = 2, lead = T, p = NULL, sigstar = NULL)
}
\arguments{
\item{x}{a number to be formatted}

\item{digits}{number of decimal places to print}
}
\description{
Internal function to format numbers for printing and plotting.
}
