% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSuggestions.R
\name{generateSuggestions}
\alias{generateSuggestions}
\title{Function to generate suggestions to initialise the genetic algorithm with.}
\usage{
generateSuggestions(items, sample_size, n_suggestions = 1000,
  required_items = NULL)
}
\arguments{
\item{items}{a character vector containing the items in the population}

\item{sample_size}{the number of items to sample}

\item{n_suggestions}{number of suggestions to initialise the GA with}

\item{required_items}{character vector of any items required to be included
in the solution}
}
\value{
a matrix suggestions to be passed to the genetic algorithm
}
\description{
Function to generate suggestions to initialise the genetic algorithm with.
}
