% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleDensity.R
\name{plotSampleDensity}
\alias{plotSampleDensity}
\title{Function to produce density curves comparing the distribution of a variable 
across different samples.}
\usage{
plotSampleDensity(items = NULL, var_name = NULL, dat = NULL,
  G = NULL, xlab = "", ylab = "", text_size = 1,
  selected_col = "black", nonselected_col = "gray", na.rm = TRUE,
  include_nonsampled = TRUE, xlim = NULL)
}
\arguments{
\item{items}{a character vector of items used to index dat}

\item{var_name}{a string contianing the name of the variable to be described}

\item{dat}{a matrix or dataframe containing describing each item in the 
population on the dimensions to be plotted}

\item{G}{a list describing each group of items}

\item{xlab}{string containing x axis label}

\item{ylab}{string containing y axis label}

\item{selected_col}{colour for plotting selected items}

\item{nonselected_col}{colour for plotting nonselected items}

\item{na.rm}{logical denoting whether to remove missing values from density
calculation}

\item{include_nonsampled}{logical denoting whether to plot nonsampled items}

\item{xlim}{an xlim vector}

\item{textsize}{cex value for descriptive text (default=1)}
}
\description{
Function to produce density curves comparing the distribution of a variable 
across different samples.
}
