% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleScatter.R
\name{plotSampleScatter}
\alias{plotSampleScatter}
\title{Function to produce scatterplots depicting bivariate distributions across
different samples.}
\usage{
plotSampleScatter(items = NULL, var_name = NULL, dat = NULL,
  G = NULL, xlim = NULL, ylim = NULL, xlab = "", ylab = "")
}
\arguments{
\item{items}{a character vector of items used to index dat}

\item{var_name}{a string contianing the name of the variable to be described}

\item{dat}{a matrix or dataframe containing describing each item in the 
population on the dimensions to be plotted}

\item{G}{a list describing each group of items}

\item{xlim}{an xlim vector}

\item{ylim}{an ylim vector}

\item{xlab}{string containing x axis label}

\item{ylab}{string containing y axis label}
}
\description{
Function to produce scatterplots depicting bivariate distributions across
different samples.
}
