% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstar.R
\name{pstar}
\alias{pstar}
\title{Internal function to create significance stars for a p value.}
\usage{
pstar(p)
}
\arguments{
\item{p}{a p value}
}
\description{
Internal function to create significance stars for a p value.
}
