% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleItems.R
\name{sampleItems}
\alias{sampleItems}
\title{Function to sample items with a genetic algorithm maximizing some measure of 
fitness (e.g., everage inter-item dissimilarity) as measured by an inter-item
distance matrix.}
\usage{
sampleItems(distance_mat, sample_size, fitness = fitnessFunction,
  lower_tri_funciton = mean, n_suggestions = 500, maxiter = 1e+06,
  required_items = NULL, seed = NULL, suggest_maxmin = FALSE)
}
\arguments{
\item{distance_mat}{a population inter-item distance matrix}

\item{sample_size}{the number of items to sample}

\item{fitness}{a fitness function to pass to ga()}

\item{lower_tri_funciton}{the specific function to assess fitness of the 
lower triangle of the sample distance matrix}

\item{n_suggestions}{number of suggestions to initialise the GA with}

\item{maxiter}{maximum number of iterations to run GA}

\item{required_items}{character vector of any items required to be included 
in the solution}

\item{seed}{seed for GA}

\item{suggest_maxmin}{logical denoting whether to include greedy search for
the items that maximise minimum inter item distance in the initial
suggestions}
}
\value{
a list containing the final items and a GA object
}
\description{
Function to sample items with a genetic algorithm maximizing some measure of 
fitness (e.g., everage inter-item dissimilarity) as measured by an inter-item
distance matrix.
}
