% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumHeight.R
\name{sumHeight}
\alias{sumHeight}
\title{Function to calculate the total dendrogram height of a distance matrix.
Inspired by Petchey, O. L., & Gaston, K. J. (2002). Functional diversity
(FD), species richness and community composition. Ecology Letters.}
\usage{
sumHeight(x)
}
\arguments{
\item{x}{a vector containing the lower triangle of a distance matrix}
}
\value{
the sum of the height of the dendrogram for the given distance matrix
}
\description{
Function to calculate the total dendrogram height of a distance matrix.
Inspired by Petchey, O. L., & Gaston, K. J. (2002). Functional diversity
(FD), species richness and community composition. Ecology Letters.
}
