% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedyMaxminSearch.R
\name{greedyMaxminSearch}
\alias{greedyMaxminSearch}
\title{Function to sample items with a greedy search maximizing minimum inter-item
distance.}
\usage{
greedyMaxminSearch(distance_mat, sample_size)
}
\arguments{
\item{distance_mat}{a population inter-item distance matrix}

\item{sample_size}{the number of items to sample}
}
\value{
a set of item names maximising minimum inter-item distance
}
\description{
Function to sample items with a greedy search maximizing minimum inter-item
distance.
}
