% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFitness.R
\name{plotFitness}
\alias{plotFitness}
\title{Function to plot the fitness of the final GA result, relative to the randomly
generated suggestions.}
\usage{
plotFitness(stim_sample, sample_size, distance_mat,
  fitness = fitnessFunction)
}
\arguments{
\item{stim_sample}{a list containing the final items and a GA object}

\item{sample_size}{the number of items to sample}

\item{distance_mat}{a population inter-item distance matrix}

\item{fitness}{the fitness function passed to ga()}
}
\description{
Function to plot the fitness of the final GA result, relative to the randomly
generated suggestions.
}
