% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleScatter.R
\name{plotSampleScatter}
\alias{plotSampleScatter}
\title{Function to produce scatterplots depicting bivariate distributions across
different samples.}
\usage{
plotSampleScatter(items = NULL, var_name = NULL, dat = NULL,
  G = NULL, xlim = NULL, ylim = NULL, xlab = "", ylab = "",
  selected_col = "black", nonselected_col = "gray",
  nonselected_alpha = 0.1)
}
\arguments{
\item{items}{a character vector of items used to index dat}

\item{var_name}{a string contianing the name of the variable to be described}

\item{dat}{a matrix or dataframe containing describing each item in the 
population on the dimensions to be plotted}

\item{G}{a list describing each group of items}

\item{xlim}{an xlim vector}

\item{ylim}{an ylim vector}

\item{xlab}{string containing x axis label}

\item{ylab}{string containing y axis label}

\item{selected_col}{colour for plotting selected items}

\item{nonselected_col}{colour for plotting nonselected items}

\item{nonselected_alpha}{transparency for scatterplot points for nonselected items}
}
\description{
Function to produce scatterplots depicting bivariate distributions across
different samples.
}
